feature(shiny): Attempted to Add Shiny
It looks like the shiny plugin does not work with ease, and it might not be
worth the struggle to add.

Cleaned up ``logistic.qmd``.
Added ``penguin.qmd`` to see if shiny would work with ``quarto preview``
or ``quarto preview ./posts/penguin.qmd``. This did not work as shown in
the documentation.
Added ``jupyter-cache`` and ``shinylive`` as dependencies.